Refactor all components' css files into one directory
