+++ src/styles/quote.css
